Various fixes for Windows
Added a cast, removed referneces to PAC, fixed up problematic tokens in heimdal

+++ vmdir/server/middle-layer/krb.c
@@ -202,7 +202,7 @@ VmDirKrbUPNKeySet(
             /* Decode existing key to get kvno value */
             dwError = VmDirKeySetGetKvno(
                           pAttrUPN->vals[0].lberbv.bv_val,
-                          pAttrUPN->vals[0].lberbv.bv_len,
+                          (DWORD) pAttrUPN->vals[0].lberbv.bv_len,
                           &kvno);
         }
 
@@ -228,7 +228,7 @@ VmDirKrbUPNKeySet(
             /* Decode existing key to get kvno value */
             dwError = VmDirKeySetGetKvno(
                           pOldKeyBlob->vals[0].lberbv.bv_val,
-                          pOldKeyBlob->vals[0].lberbv.bv_len,
+                          (DWORD) pOldKeyBlob->vals[0].lberbv.bv_len,
                           &kvno);
             BAIL_ON_VMDIR_ERROR(dwError);
             kvno++;
